Warning Fixes (#844)
* chore(warnings): progress commit - clear C4189, C4456, C4457, C4458, C4101, C4065, C4389, C4505, C4701, C4702 categories

Renames inner shadowing locals, removes unused/unreachable code, and flattens single-default switches. No public API changes; build clean, all unit tests pass. Adds WARNINGS.md tracking the sweep.


* chore(warnings): progress commit - clear C4018, C4267 categories

Promotes int loop counters to uint32_t to match unsigned bounds (6 sites) and adds explicit static_cast<T>(...) for size_t -> uint32_t/int/uint8_t narrowings (30 sites). No public API changes; build clean, all categories of these codes are at zero. Updates WARNINGS.md to reflect live state.


* chore(warnings): progress commit - clear C4018, C4189, C4245, C4267, C4456 categories

Replaces -1 sentinels assigned to unsigned types with explicit static_cast<DestType>(-1) (177 sites total across model + controller + view + tests), promotes a few remaining int loop counters to uint32_t, removes verified-dead locals, and renames inner shadowing locals where the inner/outer types differ. Also rewrites INVALID_BRANCH_LEVEL to drop a stray trailing ; and uses EventManager::InvalidSubscriptionToken instead of a magic -1 in TimelineArrow. No public API changes; build clean, all five categories at zero. Updates WARNINGS.md to reflect live state.


* chore(warnings): progress commit - clear C4267 + trivial C4996 sites

Adds explicit static_cast<T>(...) on size_t -> uint16_t/uint32_t/int narrowings (35 sites across controller + view/widgets/compute), replaces vsprintf with vsnprintf in 3 test/demo PrintHeader helpers, and renames POSIX getch to _getch in the demo main. C4267 category at zero; C4996 down to 9 (the 4 strncpy + 5 getenv sites are deferred per WARNINGS.md §4.3 - they need API-contract decisions and a portable wrapper, not auto-sweep). No public API changes; build clean, all 4 test suites pass (173M+ assertions). Updates WARNINGS.md to reflect live state.


* chore(warnings): progress commit - clear C4100 category

Comments out unreferenced parameter names (Type name -> Type /*name*/) across 30 files in model + controller + view + tests. 165 unique sites, 481 raw warnings (the headers are included by many translation units, so MSVC re-emits per-TU). All sites are intentional: virtual base no-op defaults that derived classes override-and-use, or "not supported here" overrides that exist for interface conformance and assert. No runtime behavior change - the parameters are still in the signature, ABI is unaffected, and any code that did reference the parameter would have failed to compile (none did, since these are exactly the names the compiler told us are unreferenced). C4100 category at zero; only C4244 (187) and the deferred C4996 (9) remain. Build clean, all five test suites pass (173M+ assertions). Updates WARNINGS.md to reflect live state.


* chore(warnings): progress commit - clear C4244 + C4996 categories

Adds explicit static_cast<DestType>(...) on lossy numeric conversions (131 unique sites / 187 raw across ~30 files), retypes a few loop counters to collapse clusters, and wraps std::transform(s, ::toupper/::tolower) calls in [](unsigned char) -> char lambdas to silence the int -> char narrowing in <algorithm>. Replaces the 4 remaining strncpy sites with equivalent-semantics memcpy and the 5 getenv sites with a new safe_getenv() -> std::string helper using _dupenv_s on MSVC. C4244 and C4996 both at zero; all 16 categories from this audit are now cleared, first-party warnings = 0. No public API changes; clean rebuild reports 0 errors / 0 warnings, all 6 ctest suites pass. Updates WARNINGS.md to reflect live state.


* chore: remove WARNINGS.md


* chore(warnings): clear C4100 in src/model and src/{controller,view}, plus INVALID_INDEX sentinel

- Replace `/*param*/` unused-parameter markers with `(void) param;`
  across controller, model (datamodel + database), tests, and view layers.
  Excludes src/model/src/database/rocprofvis_db_compute.h.
- Introduce shared INVALID_INDEX (0xFFFFFFFFu) and INVALID_INDEX_64
  (0xFFFFFFFFFFFFFFFFull) macros in rocprofvis_shared_types.h and
  replace ``static_cast<uintNN_t>(-1)`` sentinel literals throughout
  controller, model, and view. EventManager::InvalidSubscriptionToken
  now derives from INVALID_INDEX_64.
- Remove redundant Sample no-op overrides (GetString, SetUInt64,
  SetObject, SetString) that only forwarded to UnhandledProperty;
  the base Handle implementation already provides the same behavior.
- Drop unused ``std::vector<std::string> queries`` lambda parameter
  from the ExecuteTransaction worker in rocprofvis_db_compute.cpp and
  rocprofvis_db_rocprof.cpp::CreateIndexes; the lambda already
  captures ``vec``, so the parameter forced a wasted per-thread copy.

No behavior change. Incremental build: 0 new warnings, 0 errors.


* refactor: address PR review feedback on warning cleanup

- table_compute: drop redundant static_cast<uint64_t> around
  INVALID_INDEX_64 (the ull literal is already uint64_t).
- interface_types.h: remove stray comment about where INVALID_INDEX
  is defined.
- timeline_view, track_item: initialize SubscriptionToken members with
  EventManager::InvalidSubscriptionToken instead of the raw
  INVALID_INDEX_64, and drop the now-unused rocprofvis_shared_types.h
  include each file no longer needs.
- utils: move safe_getenv from an anonymous namespace into
  RocProfVis::View (qualifying the one call from the file-local
  display_looks_forwarded helper).
- track: serve kRPVControllerTrackHistogramBucketDensityIndexed from
  GetUInt64 instead of GetDouble (it is a uint64 count, not a double);
  update the two callers (data_provider, controller system tests) to
  rocprofvis_controller_get_uint64.
- table_processor: use uint32_t for the aggregation thread/row locals,
  and declare query offset/limit as uint32_t (atoll -> atol). Removes
  the size_t/int comparison warnings and the redundant row_index casts.


* fix: use portable getchar() in datamodel test to fix Linux build


* fix: avoid UB in UpdateSizeLimit when total weighted size is zero


* fix: resolve CodeQL alerts - sanitize env-derived config path and drop no-op const returns

+++ src/model/src/database/rocprofvis_db_query_builder.cpp
@@ -324,11 +324,11 @@ namespace DataModel
     }
 
     const uint8_t Builder::TypeEnumToInt(const char* type, std::vector<std::string>& lookup) {
-        for (int i=0; i < lookup.size(); i++)
+        for (size_t i = 0; i < lookup.size(); i++)
         {
             if (lookup[i] == type)
             {
-                return i;
+                return static_cast<uint8_t>(i);
             }
         }
         return 0;
