+++ src/model/src/datamodel/rocprofvis_dm_track.cpp
@@ -177,7 +177,7 @@ rocprofvis_dm_charptr_t  Track::CategoryString(){
 
 rocprofvis_dm_value_t Track::GetHistogramBucketValueAt(size_t index)
 { 
-    auto it = m_track_params->histogram.find(index);
+    auto it = m_track_params->histogram.find(static_cast<uint32_t>(index));
     if (it == m_track_params->histogram.end())
     {
         return 0;
@@ -189,7 +189,7 @@ rocprofvis_dm_value_t Track::GetHistogramBucketValueAt(size_t index)
 }
 
 uint64_t  Track::GetHistogramBucketNumEventsAt(size_t index) {
-    auto it = m_track_params->histogram.find(index);
+    auto it = m_track_params->histogram.find(static_cast<uint32_t>(index));
     if (it == m_track_params->histogram.end())
     {
         return 0;
